Separate module for command line tool
The command line tool uses these modules, which are very large:

    github.com/spf13/cobra
    github.com/spf13/viper

These are not used by "oviewer", but if the repository only has one module, then
the command line requirements will also be "oviewer" requirements. Ideally,
"oviewer" would be at the top level like this:

    github.com/noborus/ov (library)
    github.com/noborus/ov/ov (command line tool)

but the change as I made it was the least disruptive. I will let the owner
decide if further changes should be made.