Before and after also enable WriteOriginal
Enables writeOriginal to be output by specifying the lines
before and after the current line.
Writing the settings in ~/.ov.yaml is similar to the behavior
of `less -X`.

BeforeWriteOriginal: 1000

Added cli options write-exit-before(b) and write-exit-after(a)
(cli option takes precedence).

If the value of write-exit-after is set to 0,
it will be the current screen height.

Added CurrentLN() to the document to return the number of current line.

Added Export() to output the specified line(range) from the document.

Fix #48.
A discussion of this fix was made in #146. Thank you @aranjan7.