Default to using XDG config path; fallback to home directory.
The default configuration file name is changed to `config.yaml` and the
default path is changed to `$XDG_CONFIG_HOME/ov/config.yaml`. A fallback
ensures that users can continue to use the original file name and path.
Also update references to the config file name in the documentation and code.