Add filter search and related packages
Add filter search feature to search documents by filter.
The 'biomap' package, which provides a bidirectional map,
is also added to support this feature.

The filter function filters only the lines containing word
in the current document and creates a new document.

Link the line numbers of the new document
to the line numbers of the original document.