Changed the method of implementing section-header
The displayed lines have been changed so that they are not affected
by section headers.
However, if there are lines that should be displayed, they will be displayed
by avoiding the section headers.
Also, if necessary, the search will start from the previous search line.
Changed to display the mark style higher than the section-Header style.