Change the range to csiParamStart/csiParamEnd
Change the parameter that was judged only by FinalyByte
to a range judgment of start and end.
Which one to judge is implementation-dependent (terminal emulator
specifies the start range, and `less` specifies only FinalByte).
The parameter will be changed to judge the start range so
that the display will not be corrupted.

Clarify Escape Sequence-> CSI -> SGR by splitting the function.

UnderLine parameters are now functions, and empty parameters have been fixed.